added loading message while average GPA is being computed, switched NaN to N/A in popover, refactored grade statistic calculation code in Pie.tsx and Chart.tsx to gradeDist.ts under helpers/
